kern: move clk_intr_event to clock.[ch]
clk_intr_event is a clock interrupt, not a generalized interrupt.  As
such everything should be in clock.h/kern_clock.c instead of
interrupt.h/kern_intr.c.
